Add unity.config into sample code
