+++ Medidata.Cloud.ExcelLoader/CellTypeValueConverterFactory.cs
@@ -5,27 +5,32 @@ using Medidata.Cloud.ExcelLoader.CellTypeConverters;
 
 namespace Medidata.Cloud.ExcelLoader
 {
-    public class CellTypeValueConverterFactory: ICellTypeValueConverterFactory
+    public class CellTypeValueConverterFactory : ICellTypeValueConverterFactory
     {
         private readonly IDictionary<Type, ICellTypeValueConverter> _converters;
 
-        public  CellTypeValueConverterFactory(params ICellTypeValueConverter[] converters)
+        public CellTypeValueConverterFactory() : this(null) {}
+
+        public CellTypeValueConverterFactory(params ICellTypeValueConverter[] converters)
         {
-            _converters = new ICellTypeValueConverter[]
-                          {
-                              new BooleanConverter(),
-                              new NullableBooleanConverter(),
-                              new DateTimeConverter(),
-                              new NullableDateTimeConverter(),
-                              new DoubleConverter(),
-                              new FloatConverter(),
-                              new DecimalConverter(),
-                              new IntConverter(),
-                              new LongConverter(),
-                              new StringConverter()
-                          }
-                          .Concat(converters)
-                          .ToDictionary(x => x.CSharpType, x => x);
+            IEnumerable<ICellTypeValueConverter> defaultConverters = new ICellTypeValueConverter[]
+                                                                     {
+                                                                         new BooleanConverter(),
+                                                                         new NullableBooleanConverter(),
+                                                                         new DateTimeConverter(),
+                                                                         new NullableDateTimeConverter(),
+                                                                         new DoubleConverter(),
+                                                                         new FloatConverter(),
+                                                                         new DecimalConverter(),
+                                                                         new IntConverter(),
+                                                                         new LongConverter(),
+                                                                         new StringConverter()
+                                                                     };
+            if (converters != null)
+            {
+                defaultConverters = defaultConverters.Concat(converters);
+            }
+            _converters = defaultConverters.ToDictionary(x => x.CSharpType, x => x);
         }
 
         public ICellTypeValueConverter Produce(Type type)
