+++ Medidata.Cloud.ExcelLoader/ExcelParser.cs
@@ -1,10 +1,8 @@
-using System;
 using System.Collections.Generic;
 using System.Dynamic;
 using System.IO;
 using DocumentFormat.OpenXml.Packaging;
 using Medidata.Cloud.ExcelLoader.Helpers;
-using Medidata.Cloud.ExcelLoader.SheetDefinitions;
 
 namespace Medidata.Cloud.ExcelLoader
 {
