--- Medidata.Cloud.ExcelLoader/ISheetDecorator.cs
@@ -1,9 +1,0 @@
-using OfficeOpenXml;
-
-namespace Medidata.Cloud.ExcelLoader
-{
-    public interface ISheetDecorator
-    {
-        void Decorate(ExcelWorksheet sheet);
-    }
-}
