+++ Medidata.Cloud.ExcelLoader/Medidata.Cloud.ExcelLoader.csproj
@@ -42,6 +42,7 @@
   <ItemGroup>
     <Compile Include="Helpers\IExtraProperty.cs" />
     <Compile Include="ICellTypeValueConverterFactory.cs" />
+    <Compile Include="IExcelLoader.cs" />
     <Compile Include="SheetDecorators\ExpandoObjectSheetDecorator.cs" />
     <Compile Include="SheetDefinitions\ColumnHeaderNameAttribute.cs" />
     <Compile Include="SheetDefinitions\ColumnIngoredAttribute.cs" />
@@ -90,6 +91,7 @@
     <Compile Include="SheetDefinitions\SheetModelExtensions.cs" />
     <Compile Include="SheetDefinitions\SheetNameAttribute.cs" />
     <Compile Include="SheetParser.cs" />
+    <Compile Include="ExcelLoader.cs" />
     <Compile Include="Validations\DefaultSequentialRuleValidator.cs" />
     <Compile Include="Validations\IValidationError.cs" />
     <Compile Include="Validations\IValidationMessage.cs" />
