+++ Medidata.Rave.Tsdv.Loader.Sample/packages.config
@@ -1,8 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 
 <packages>
-  <package id="AutoFixture" version="3.37.3" targetFramework="net4" />
-  <package id="AutoFixture.AutoRhinoMocks" version="3.37.3" targetFramework="net4" />
   <package id="Medidata.Interfaces" version="0.1.3" targetFramework="net40" />
-  <package id="RhinoMocks" version="3.6.1" targetFramework="net4" />
+  <package id="Unity" version="2.1.505.2" targetFramework="net4" />
 </packages>
