+++ Medidata.Rave.Tsdv.Loader.Sample/unity.config
@@ -1,0 +1,63 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+
+<unity xmlns="http://schemas.microsoft.com/practices/2010/unity">
+  <container>
+    <register type="Medidata.Interfaces.Localization.ILocalization, Medidata.Interfaces"
+              mapTo="Medidata.Rave.Tsdv.Loader.Sample.Localization, Medidata.Rave.Tsdv.Loader.Sample" />
+
+    <register type="Medidata.Cloud.ExcelLoader.ISheetBuilderDecorator, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.SheetDecorators.TextStyleSheetDecorator, Medidata.Cloud.ExcelLoader"
+              name="textStyleDecorator">
+      <constructor>
+        <param name="styleName" value="Normal" />
+      </constructor>
+    </register>
+
+    <register type="Medidata.Cloud.ExcelLoader.ISheetBuilderDecorator, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.SheetDecorators.HeaderStyleSheetDecorator, Medidata.Cloud.ExcelLoader"
+              name="headerStyleDecorator">
+      <constructor>
+        <param name="styleName" value="Output" />
+      </constructor>
+    </register>
+
+    <register type="Medidata.Cloud.ExcelLoader.ICellTypeValueConverterFactory, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.CellTypeValueConverterFactory, Medidata.Cloud.ExcelLoader"
+              name="converterFactory">
+      <lifetime type="singleton" />
+    </register>
+    <register type="Medidata.Cloud.ExcelLoader.ISheetBuilder, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.SheetBuilder, Medidata.Cloud.ExcelLoader">
+      <constructor>
+        <param name="converterFactory" dependencyName="converterFactory" />
+        <param name="decorators">
+          <array>
+            <dependency
+              type="Medidata.Cloud.ExcelLoader.SheetDecorators.HeaderSheetDecorator, Medidata.Cloud.ExcelLoader" />
+            <dependency
+              type="Medidata.Cloud.ExcelLoader.SheetDecorators.AutoFilterSheetDecorator, Medidata.Cloud.ExcelLoader" />
+            <dependency name="textStyleDecorator" />
+            <dependency name="headerStyleDecorator" />
+            <dependency type="Medidata.Rave.Tsdv.Loader.TranslateHeaderDecorator, Medidata.Rave.Tsdv.Loader" />
+            <dependency
+              type="Medidata.Cloud.ExcelLoader.SheetDecorators.AutoFitWidthSheetDecorator, Medidata.Cloud.ExcelLoader" />
+          </array>
+        </param>
+
+      </constructor>
+    </register>
+    <register type="Medidata.Cloud.ExcelLoader.ISheetParser, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.SheetParser, Medidata.Cloud.ExcelLoader">
+      <constructor>
+        <param name="converterFactory" dependencyName="converterFactory" />
+      </constructor>
+    </register>
+    <register type="Medidata.Cloud.ExcelLoader.IExcelBuilder, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Rave.Tsdv.Loader.CoveredExcelBuilder, Medidata.Rave.Tsdv.Loader" />
+    <register type="Medidata.Cloud.ExcelLoader.IExcelParser, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Cloud.ExcelLoader.ExcelParser, Medidata.Cloud.ExcelLoader" />
+
+    <register type="Medidata.Cloud.ExcelLoader.IExcelLoader, Medidata.Cloud.ExcelLoader"
+              mapTo="Medidata.Rave.Tsdv.Loader.TsdvExcelLoader, Medidata.Rave.Tsdv.Loader" />
+  </container>
+</unity>
