--- Medidata.Rave.Tsdv.Loader/ITsdvReportLoader.cs
@@ -1,10 +1,0 @@
-﻿using System.IO;
-
-namespace Medidata.Rave.Tsdv.Loader
-{
-    public interface ITsdvReportLoader
-    {
-        void Save(Stream outStream);
-        void Load(Stream source);
-    }
-}
