+++ Medidata.Rave.Tsdv.Loader/Medidata.Rave.Tsdv.Loader.csproj
@@ -43,10 +43,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ITsdvReportLoader.cs" />
     <Compile Include="SheetDefinitions\v1\BlockPlan.cs" />
-    <Compile Include="SheetDefinitions\TranslateHeaderDecorator.cs" />
-    <Compile Include="SheetDefinitions\TsdvReportLoader.cs" />
+    <Compile Include="TranslateHeaderDecorator.cs" />
+    <Compile Include="TsdvExcelLoader.cs" />
     <Compile Include="CoveredExcelBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resource.Designer.cs">
